# Optimize `Range` with `Integer`