`rb_obj_is_kind_of(v, rb_cNumeric)` is a check that often appears in the source code.
In particular, when `v` is `Integer`, it is faster to use `rb_integer_type_p`.
This pull request defines `is_numeric_p` which combines `rb_integer_type_p` and `rb_obj_is_kind_of` to more quickly determine that an `Integer` is `Numeric`.
This is especially effective for `Range#size`.

# benchmark

```
prelude: |
  r_1 = 1..1
  r_1k = 1..1000
  r_1m = 1..1000000
  r_str = 'a'..'z'
benchmark:
  'int 1': r_1.size
  'int 1K': r_1k.size
  'int 1M': r_1m.size
  string: r_str.size
```

## Iteration per second (i/s)

|        |master|PR|
|:-------|----------------------------------:|------------------------:|
|int 1   |                            34.897M|                  43.074M|
|        |                                  -|                    1.23x|
|int 1K  |                            35.023M|                  42.940M|
|        |                                  -|                    1.23x|
|int 1M  |                            34.855M|                  42.560M|
|        |                                  -|                    1.22x|
|string  |                            46.039M|                  45.110M|
|        |                              1.02x|                        -|